requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
aiosqlite==0.20.0
aiofiles==24.1.0
orjson==3.10.12
structlog==24.4.0
diskcache==5.6.3
chromadb==0.5.23
posthog==3.5.0
numpy==1.26.4
python-multipart==0.0.9
pytest==8.3.4
pytest-asyncio==0.25.0
TTS==0.22.0
